requirements.txt
nltk
sentence-transformers
faiss-cpu